RagSystem/requirements.txt
langchain==0.1.16
langchain-community==0.0.34
langchain-openai==0.1.3
pypdf==4.2.0
chromadb==0.4.24
python-dotenv==1.0.1
tiktoken==0.6.0
